Refactors example includes for improved readability
Replaces `highlight` and `include_relative` tags with a single `example` tag in the documentation.

This simplifies the syntax for including code examples and improves the overall readability of the markdown files. It also fixes a minor rendering issue in the example plugin related to parsing options with quotes and commas.

Increases the border size of the state selector example for better visual clarity.